backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
pytest-pythonpath==0.7.3
django-filter==22.1
djoser==2.1.0
python-dotenv==0.21.1
Pillow==8.3.1
drf-base64==2.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
isort == 5.11.5
black==23.3.0
